fix #6533 access global type filter in filter bar of cache list
