+++ main/src/cgeo/geocaching/ui/dialog/Dialogs.java
@@ -1,7 +1,11 @@
 package cgeo.geocaching.ui.dialog;
 
 import cgeo.geocaching.CgeoApplication;
+import cgeo.geocaching.R;
+import cgeo.geocaching.enumerations.CacheType;
+import cgeo.geocaching.settings.Settings;
 import cgeo.geocaching.utils.ImageUtils;
+import cgeo.geocaching.utils.TextUtils;
 import cgeo.geocaching.utils.functions.Action1;
 
 import android.app.Activity;
@@ -25,6 +29,10 @@ import android.widget.EditText;
 import android.widget.ListAdapter;
 import android.widget.TextView;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 import butterknife.ButterKnife;
@@ -500,4 +508,48 @@ public final class Dialogs {
             dialog.dismiss();
         }
     }
+
+    public static void selectGlobalTypeFilter(final Activity activity, final Action1<CacheType> okayListener) {
+        final List<CacheType> cacheTypes = new ArrayList<>();
+
+        //first add the most used types
+        cacheTypes.add(CacheType.ALL);
+        cacheTypes.add(CacheType.TRADITIONAL);
+        cacheTypes.add(CacheType.MULTI);
+        cacheTypes.add(CacheType.MYSTERY);
+
+        // then add all other cache types sorted alphabetically
+        final List<CacheType> sorted = new ArrayList<>(Arrays.asList(CacheType.values()));
+        sorted.removeAll(cacheTypes);
+        Collections.sort(sorted, new Comparator<CacheType>() {
+            @Override
+            public int compare(final CacheType left, final CacheType right) {
+                return TextUtils.COLLATOR.compare(left.getL10n(), right.getL10n());
+            }
+        });
+        cacheTypes.addAll(sorted);
+
+        final int checkedItem = Math.max(0, cacheTypes.indexOf(Settings.getCacheType()));
+
+        final String[] items = new String[cacheTypes.size()];
+        for (int i = 0; i < cacheTypes.size(); i++) {
+            items[i] = cacheTypes.get(i).getL10n();
+        }
+
+        final Builder builder = new AlertDialog.Builder(activity);
+        builder.setTitle(R.string.menu_filter);
+        builder.setSingleChoiceItems(items, checkedItem, new DialogInterface.OnClickListener() {
+
+            @Override
+            public void onClick(final DialogInterface dialog, final int position) {
+                final CacheType cacheType = cacheTypes.get(position);
+                Settings.setCacheType(cacheType);
+                okayListener.call(cacheType);
+                dialog.dismiss();
+            }
+
+        });
+        builder.create().show();
+    }
+
 }
